More tidyup
